move core functions definitions to core file + define instance method add()
